Move away from nested tox.ini
