+++ charms/worker/k8s/pyproject.toml
@@ -34,6 +34,105 @@ unit = [
     "pytest>=8.3.4",
     "pytest-sugar"
 ]
+static = ["bandit"]
+update = ["pyyaml"]
+lint = [
+    "mypy",
+    "ruff",
+    "pydantic[mypy]",
+    "types-PyYAML"
+]
+
+[tool.bandit]
+exclude_dirs = ["lib/"]
+skips = ["B404","B603"]
+[tool.bandit.assert_used]
+skips = ["*/*test.py", "*/test_*.py", "*tests/*.py"]
+
+# Testing tools configuration
+[tool.coverage.run]
+branch = true
+
+[tool.coverage.report]
+show_missing = true
+
+[tool.coverage.xml]
+output = "coverage.xml"
+
+
+[tool.isort]
+line_length = 99
+profile = "black"
+
+# Linting tools configuration
+[tool.flake8]
+max-line-length = 99
+max-doc-length = 99
+max-complexity = 10
+exclude = [".git", "__pycache__", ".tox", "build", "dist", "*.egg_info", "venv"]
+select = ["E", "W", "F", "C", "N", "R", "D", "H"]
+# Ignore W503, E501 because using black creates errors with this
+# Ignore D107 Missing docstring in __init__
+# Ignore N805 first argument should be named self. Pydantic validators do not comply.
+ignore = ["W503", "E501", "D107", "N805"]
+
+# D100, D101, D102, D103: Ignore missing docstrings in tests
+per-file-ignores = ["tests/*:D100,D101,D102,D103,D104,D205,D212,D415"]
+docstring-convention = "google"
+
+# Linting tools configuration
+[tool.ruff]
+line-length = 99
+exclude = ["lib/"]
+extend-exclude = ["__pycache__", "*.egg_info"]
+
+[tool.ruff.lint]
+select = ["E", "W", "F", "C", "N", "D", "I001"]
+ignore = [
+    "D203",
+    "D204",
+    "D213",
+    "D215",
+    "D400",
+    "D404",
+    "D406",
+    "D407",
+    "D408",
+    "D409",
+    "D413",
+    "D107",
+    "N805",
+#    "E503",
+]
+per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104"]}
+
+[tool.pylint]
+# Ignore too-few-public-methods due to pydantic models
+# Ignore no-self-argument due to pydantic validators
+disable = "wrong-import-order,redefined-outer-name,too-many-instance-attributes,too-few-public-methods,no-self-argument,fixme,protected-access"
+# Ignore Pydantic check: https://github.com/pydantic/pydantic/issues/1961
+extension-pkg-whitelist = "pydantic" # wokeignore:rule=whitelist
+# Modules can be bigger than 1000 lines
+max-module-lines = 1500
+
+[tool.ruff.lint.mccabe]
+max-complexity = 10
+
+[tool.codespell]
+skip = "build,lib,venv,icon.svg,.tox,.git,.mypy_cache,.ruff_cache,.coverage"
+
+[tool.mypy]
+ignore-missing-imports = true
+explicit-package-bases = true
+namespace-packages = true
+plugins = "pydantic.mypy"
+
+[tool.pylint.typecheck]
+# Ignore typechecking on pylxd manager classes
+ignored-classes = "ProfileManager,InstanceManager,NetworkManager,FieldInfo"
+
+[tool.pytest.ini_options]
+minversion = "6.0"
 
 [tool.pyright]
 extraPaths = ["src", "lib"]
