--- charms/worker/k8s/tox.ini
@@ -1,59 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-[tox]
-skipsdist=True
-skip_missing_interpreters = True
-envlist = unit
-
-[vars]
-src_path = {toxinidir}/src/
-tst_path = {toxinidir}/tests/
-lib_path = {toxinidir}/lib/charms/k8s
-scripts_path = {toxinidir}/scripts
-all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
-
-[testenv]
-setenv =
-  PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
-  PYTHONBREAKPOINT=ipdb.set_trace
-  PY_COLORS=1
-passenv =
-  PYTHONPATH
-  CHARM_BUILD_DIR
-  MODEL_SETTINGS
-
-[testenv:unit]
-runner = uv-venv-lock-runner
-description = Run unit tests
-dependency_groups = unit
-commands =
-    uv run coverage run --source={[vars]src_path},{[vars]lib_path} \
-        -m pytest --ignore={[vars]tst_path}integration -vv \
-        --basetemp={envtmpdir} \
-        --tb native -s {posargs}
-    uv run coverage report --show-missing
-
-[testenv:coverage-report]
-description = Create test coverage report
-deps =
-    coverage[toml]
-commands =
-    coverage report
-
-[testenv:coverage-xml]
-description = Create test coverage XML report
-deps =
-    coverage[xml]
-commands =
-    coverage xml
-
-[testenv:update-dashboards]
-description = Run the Grafana dashboards update script
-deps = pyyaml
-commands = python {[vars]scripts_path}/update_dashboards.py
-
-[testenv:update-alert-rules]
-description = Run the AlertManager rules update script
-deps = pyyaml
-commands = python {[vars]scripts_path}/update_alert_rules.py
