+++ tests/unit/test_base.py
@@ -33,23 +33,6 @@ from charms.k8s.v0.k8sd_api_manager import (
 )
 
 
-@pytest.fixture(params=["worker", "control-plane"])
-def harness(request):
-    """Craft a ops test harness.
-
-    Args:
-        request: pytest request object
-    """
-    meta = Path(__file__).parent / "../../charmcraft.yaml"
-    if request.param == "worker":
-        meta = Path(__file__).parent / "../../../charmcraft.yaml"
-    harness = ops.testing.Harness(K8sCharm, meta=meta.read_text())
-    harness.begin()
-    harness.charm.is_worker = request.param == "worker"
-    yield harness
-    harness.cleanup()
-
-
 @contextlib.contextmanager
 def mock_reconciler_handlers(harness):
     """Mock out reconciler handlers.
