+++ tests/unit/test_inspector.py
