+++ tests/unit/test_k8sd_api_manager.py
