+++ tests/unit/test_token_distributor.py
@@ -16,23 +16,6 @@ from charm import K8sCharm
 from literals import CLUSTER_RELATION
 
 
-@pytest.fixture(params=["worker", "control-plane"])
-def harness(request):
-    """Craft a ops test harness.
-
-    Args:
-        request: pytest request object
-    """
-    meta = Path(__file__).parent / "../../charmcraft.yaml"
-    if request.param == "worker":
-        meta = Path(__file__).parent / "../../../charmcraft.yaml"
-    harness = ops.testing.Harness(K8sCharm, meta=meta.read_text())
-    harness.begin()
-    harness.charm.is_worker = request.param == "worker"
-    yield harness
-    harness.cleanup()
-
-
 def test_request(harness):
     """Test request adds node-name."""
     harness.disable_hooks()
