+++ tox.ini
@@ -7,10 +7,20 @@ skip_missing_interpreters = True
 envlist = lint-mock, lint, unit, static, coverage-report, coverage-xml
 
 [vars]
-lib_path = {toxinidir}/charms/worker/k8s/lib
-src_path = {toxinidir}/charms/worker/k8s/src {[vars]lib_path}/charms/k8s {toxinidir}/charms/worker/k8s/scripts
-tst_path = {toxinidir}/tests/ {toxinidir}/charms/worker/k8s/tests/
-all_path = {[vars]tst_path} {[vars]src_path}
+charm_path   = {toxinidir}/charms/worker/k8s
+lib_path     = {[vars]charm_path}/lib
+src_path     = {[vars]charm_path}/src
+charm_lib    = {[vars]lib_path}/charms/k8s
+scripts_path = {[vars]charm_path}/scripts
+test_base    = {toxinidir}/tests
+int_path     = {[vars]test_base}/integration
+unit_path    = {[vars]test_base}/unit
+all_path     = {[vars]test_base} {[vars]charm_path}
+uv_sync_flags =
+    --project
+    {[vars]charm_path}
+uv_run       =  uv run --project {[vars]charm_path} --directory {[vars]charm_path}
+
 
 [testenv]
 setenv =
@@ -20,15 +30,6 @@ passenv =
   PYTHONPATH
   CHARM_BUILD_DIR
   MODEL_SETTINGS
-
-[testenv:format]
-allowlist_externals = tox
-description = Apply coding style standards to code
-deps =
-    ruff
-commands =
-    ruff format {[vars]all_path}
-    ruff check --fix {[vars]all_path}
 
 [testenv:lint-mock]
 description = Provide mock lint commands to use ruff instead
@@ -47,41 +48,66 @@ commands =
 
 [testenv:lint]
 description = Check code against coding style standards
-setenv =
-  PYTHONPATH = {envdir}{:}{[vars]lib_path}
-deps =
-    mypy
-    ruff
-    pydantic[mypy]
-    types-PyYAML
+runner = uv-venv-lock-runner
+dependency_groups = lint
+uv_sync_flags = {[vars]uv_sync_flags}
+envdir = .tox/lintenv
 commands =
-    ruff format --check {[vars]all_path}
-    ruff check {[vars]all_path}
-    mypy {[vars]all_path}
+    {[vars]uv_run} ruff format --check {[vars]all_path}
+    {[vars]uv_run} ruff check {[vars]all_path}
+    {[vars]uv_run} mypy {[vars]all_path}
+
+[testenv:format]
+description = Apply coding style standards to code
+runner = uv-venv-lock-runner
+dependency_groups = lint
+uv_sync_flags = {[vars]uv_sync_flags}
+envdir = .tox/lintenv
+commands =
+    {[vars]uv_run} ruff format {[vars]all_path}
+    {[vars]uv_run} ruff check --fix {[vars]all_path}
 
 [testenv:unit]
-allowlist_externals = tox
-deps =
-    -r test_requirements.txt
+description = Run unit tests
+runner = uv-venv-lock-runner
+dependency_groups = unit
+uv_sync_flags = {[vars]uv_sync_flags}
+envdir = .tox/unit
+setenv =
+    PYTHONPATH = {[vars]src_path}:{[vars]lib_path}
 commands =
-    tox -c {toxinidir}/charms/worker/k8s -e unit -- {posargs}
+    {[vars]uv_run} coverage run --source={[vars]src_path},{[vars]charm_lib} \
+        -m pytest {[vars]unit_path} -vv \
+        --basetemp={envtmpdir} \
+        --tb native -s {posargs}
+    {[vars]uv_run} coverage report --show-missing
 
 [testenv:coverage-report]
-allowlist_externals = tox
+description = Create test coverage report
+runner = uv-venv-lock-runner
+dependency_groups = unit
+envdir = .tox/unit
+uv_sync_flags = {[vars]uv_sync_flags}
 commands =
-    tox -c {toxinidir}/charms/worker/k8s -e coverage-report
+    {[vars]uv_run} coverage report
 
 [testenv:coverage-xml]
-allowlist_externals = tox
+description = Create test coverage report
+runner = uv-venv-lock-runner
+dependency_groups = unit
+envdir = .tox/unit
+uv_sync_flags = {[vars]uv_sync_flags}
 commands =
-    tox -c {toxinidir}/charms/worker/k8s -e coverage-xml
+    {[vars]uv_run} coverage xml
 
 [testenv:static]
 description = Run static analysis tests
-deps =
-    bandit[toml]
+runner = uv-venv-lock-runner
+dependency_groups = static
+envdir = .tox/static
+uv_sync_flags = {[vars]uv_sync_flags}
 commands =
-    bandit -c {toxinidir}/pyproject.toml -r {[vars]all_path}
+    {[vars]uv_run} bandit -c {[vars]charm_path}/pyproject.toml -r {[vars]charm_path}
 
 [testenv:{integration,integration-k8s,integration-etcd,integration-ceph,integration-upgrade}]
 description = Run integration tests
@@ -113,3 +139,21 @@ deps =
 commands =
     ; can't run lazydocs directly due to needing to run it on src/* which produces an invocation error in tox
     sh generate-src-docs.sh
+
+[testenv:update-dashboards]
+description = Run the Grafana dashboards update script
+runner = uv-venv-lock-runner
+dependency_groups = update
+envdir = .tox/update
+uv_sync_flags = {[vars]uv_sync_flags}
+commands =
+    {[vars]uv_run} --directory {[vars]charm_path} {[vars]scripts_path}/update_dashboards.py
+
+[testenv:update-alert-rules]
+description = Run the AlertManager rules update script
+runner = uv-venv-lock-runner
+dependency_groups = update
+envdir = .tox/update
+uv_sync_flags = {[vars]uv_sync_flags}
+commands =
+    {[vars]uv_run} --directory {[vars]charm_path} {[vars]scripts_path}/update_alert_rules.py
